"Refactor error handling in baidu-tts and app
Updated the error handling code in the getAccessToken() function of the baidu-tts module and in the text handler of the app module. Replaced existing `.then(err => {})` error handle code with a `.catch(err => {})` in order to catch and handle errors more effectively. Added a message in app.js to notify the user when an error occurs."