Add error check for Baidu TTS response content type
Added a conditional check in baidu-tts.js to verify if the response content-type from Baidu TTS is 'audio/mp3'. If not, an error is thrown. This is to avoid potential issues when the received response is not in the expected format. Also, improved error message formatting in app.js to be more readable by separating the general and specific error messages with a newline.