Make doSpeak function asynchronous, improve user interactivity
Updated the doSpeak function in app.js to be asynchronous. This change was important and necessary to improve user interactivity. Previously, the user was not informed when the audio file was being generated. Now, with these changes implemented, the bot sends a message to the user advising that the audio file is currently being generated, and again when the file is ready or if an error occurs.